requirements.txt
streamlit
requests
tweepy
python-dotenv
pandas